requirements_simple.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.0
sentence-transformers>=2.6.0
transformers>=4.35.0
torch>=2.1.0
tiktoken>=0.5.0
